# openapi.yaml
openapi: 3.0.3
info:
  title: HomeFlip API
  description: Production-ready house flipping and rental property management API with advanced contractor verification
  version: 1.0.0
  contact:
    name: HomeFlip Support
    email: support@homeflip.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.homeflip.com/v1
    description: Production server
  - url: https://staging-api.homeflip.com/v1
    description: Staging server
  - url: http://localhost:5000/api
    description: Development server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /auth/register:
    post:
      tags: [Authentication]
      summary: Register new user
      description: Create a new user account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags: [Authentication]
      summary: User login
      description: Authenticate user and return access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/me:
    get:
      tags: [Authentication]
      summary: Get current user
      description: Get current authenticated user information
      responses:
        '200':
          description: User information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Properties
...
  /admin/compute-scores:
    post:
      tags: [Admin]
      summary: Compute contractor scores
      description: Trigger recomputation of all contractor scores
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Scores computed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComputeScoresResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/audit:
    get:
      tags: [Admin]
      summary: Get audit logs
      description: Get system audit logs with filtering
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: object
          in: query
          schema:
            type: string
        - name: action
          in: query
          schema:
            type: string
            enum: [CREATE, UPDATE, DELETE, VIEW, LOGIN, LOGOUT, EXPORT, IMPORT]
        - name: actorId
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Audit logs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAuditLogs'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number (1-based)
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: pageSize
      in: query
      description: Number of items per page
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Error code
            message:
              type: string
              description: Error message
            details:
              type: object
              description: Additional error details
            requestId:
              type: string
              description: Request ID for tracking
...
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
          description: JWT access token
        user:
          $ref: '#/components/schemas/User'
    RegisterRequest:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          pattern: '^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)'
        company:
          type: string
          maxLength: 100
        phone:
          type: string
          pattern: '^\\+?[1-9]\\d{1,14}$'
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    User:
      type: object
      required: [id, name, email, role, isActive, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        company:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        role:
          type: string
          enum: [ADMIN, USER, CONTRACTOR, STAFF]
        isActive:
          type: boolean
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
        timezone:
          type: string
          default: UTC
        language:
          type: string
          default: en
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Pagination:
      type: object
      required: [page, pageSize, total, totalPages]
      properties:
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
...
    ComputeScoresResponse:
      type: object
      required: [message, results, totalProcessed]
      properties:
        message:
          type: string
        results:
          type: array
          items:
            type: object
            properties:
              contractorId:
                type: string
              name:
                type: string
              score:
                type: integer
              grade:
                type: string
        totalProcessed:
          type: integer
    PaginatedAuditLogs:
      type: object
      required: [auditLogs, pagination]
      properties:
        auditLogs:
          type: array
          items:
            $ref: '#/components/schemas/AuditLog'
        pagination:
          $ref: '#/components/schemas/Pagination'
    AuditLog:
      type: object
      required: [id, action, targetType, targetId, createdAt]
      properties:
        id:
          type: string
        actorId:
          type: string
          nullable: true
        action:
          type: string
          enum: [CREATE, UPDATE, DELETE, VIEW, LOGIN, LOGOUT, EXPORT, IMPORT]
        targetType:
          type: string
        targetId:
          type: string
        meta:
          type: object
        ipAddress:
          type: string
          nullable: true
        userAgent:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        actor:
          type: object
          nullable: true
          properties:
            id:
              type: string
            name:
              type: string
            email:
              type: string
...